+++ src/components/Layout/Footer.tsx
@@ -1,5 +1,8 @@
+import { css } from '@emotion/react';
+
 import { GithubIcon } from '@/assets/svgs';
-import { footerWrapperStyle } from './Footer.styles';
+import { COLORS } from '@/styles/colors';
+import { SIZES } from '@/styles/sizes';
 
 interface Props {
   githubUsername: string;
@@ -16,3 +19,31 @@ export const Footer = ({ githubUsername, bioGithubLink }: Props) => {
     </footer>
   );
 };
+
+const footerWrapperStyle = css({
+  width: '100vw',
+  display: 'flex',
+  flexDirection: 'row',
+  alignItems: 'center',
+  justifyContent: 'center',
+  padding: 50,
+
+  span: {
+    color: COLORS.GRAY9,
+    marginRight: 10,
+    fontSize: SIZES.$FONT_M,
+    userSelect: 'none',
+    cursor: 'default',
+  },
+
+  a: {
+    padding: 0,
+    margin: 0,
+    width: 20,
+    height: 20,
+  },
+
+  svg: {
+    fill: COLORS.GRAY9,
+  },
+});
