+++ src/components/Layout/index.tsx
@@ -1,10 +1,12 @@
-import { Header } from './Header';
+import { css } from '@emotion/react';
 
 import { ProfileType } from '@/types/profileData';
+import { Header } from './Header';
 import { Footer } from './Footer';
 import { HeadMetaData } from './HeadMetaData';
 
-import { mainWrapperStyle } from './Layout.styles';
+import { COLORS } from '@/styles/colors';
+import { getResponsiveAfter } from '@/styles/getResponsiveBreakpoint';
 
 interface Props {
   children: JSX.Element;
@@ -23,3 +25,21 @@ export const Layout = ({ children, profileData }: Props) => {
     </>
   );
 };
+
+const mainWrapperStyle = css({
+  padding: '0 25px',
+
+  [getResponsiveAfter('SD')]: {
+    padding: '0 100px',
+  },
+
+  '> div': {
+    backgroundColor: COLORS.WHITE,
+    padding: 25,
+    borderRadius: 5,
+
+    [getResponsiveAfter('SD')]: {
+      padding: 50,
+    },
+  },
+});
